requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.1
huggingface-hub==0.11.1
nltk==3.8.1
numpy==1.24.1
python-dotenv==0.21.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
sentence-transformers==2.2.2
transformers==4.25.1